% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Takes the euclidian algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{shall}{be a numeric value and a scalar}

\item{shall}{be a numeric value and a scalar}
}
\description{
This is a ahappy moment :D
}
\examples{
\dontrun{
# Dont run this
}
power_x(x=10)
}
