\name{First-package}
\alias{First-package}
\alias{First}
\docType{package}
\title{
\packageTitle{First}
}
\description{
\packageDescription{First}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{First}
\packageIndices{First}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{First}

Maintainer: \packageMaintainer{First}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
